polish(pu): set output_norm=True for all MLP model
