+++ lzero/model/muzero_model_mlp.py
@@ -13,26 +13,27 @@ from .utils import renormalize, get_params_mean, get_dynamic_mean, get_reward_me
 class MuZeroModelMLP(nn.Module):
 
     def __init__(
-            self,
-            observation_shape: int = 2,
-            action_space_size: int = 6,
-            latent_state_dim: int = 256,
-            fc_reward_layers: SequenceType = [32],
-            fc_value_layers: SequenceType = [32],
-            fc_policy_layers: SequenceType = [32],
-            reward_support_size: int = 601,
-            value_support_size: int = 601,
-            proj_hid: int = 1024,
-            proj_out: int = 1024,
-            pred_hid: int = 512,
-            pred_out: int = 1024,
-            self_supervised_learning_loss: bool = False,
-            categorical_distribution: bool = True,
-            activation: Optional[nn.Module] = nn.ReLU(inplace=True),
-            last_linear_layer_init_zero: bool = True,
-            state_norm: bool = False,
-            *args,
-            **kwargs
+        self,
+        observation_shape: int = 2,
+        action_space_size: int = 6,
+        latent_state_dim: int = 256,
+        fc_reward_layers: SequenceType = [32],
+        fc_value_layers: SequenceType = [32],
+        fc_policy_layers: SequenceType = [32],
+        reward_support_size: int = 601,
+        value_support_size: int = 601,
+        proj_hid: int = 1024,
+        proj_out: int = 1024,
+        pred_hid: int = 512,
+        pred_out: int = 1024,
+        self_supervised_learning_loss: bool = False,
+        categorical_distribution: bool = True,
+        activation: Optional[nn.Module] = nn.ReLU(inplace=True),
+        last_linear_layer_init_zero: bool = True,
+        state_norm: bool = False,
+        discrete_action_encoding_type: str = 'one_hot',
+        *args,
+        **kwargs
     ):
         """
         Overview:
@@ -60,6 +61,7 @@ class MuZeroModelMLP(nn.Module):
                 operation to speedup, e.g. ReLU(inplace=True).
             - last_linear_layer_init_zero (:obj:`bool`): Whether to use zero initialization for the last layer of value/policy mlp, default set it to True.
             - state_norm (:obj:`bool`): Whether to use normalization for latent states, default set it to True.
+            - discrete_action_encoding_type (:obj:`str`): The encoding type of discrete action, which can be 'one_hot' or 'not_one_hot'.
         """
         super(MuZeroModelMLP, self).__init__()
         self.categorical_distribution = categorical_distribution
@@ -71,8 +73,20 @@ class MuZeroModelMLP(nn.Module):
             self.value_support_size = value_support_size
 
         self.action_space_size = action_space_size
-        # for discrete action space, we use one-hot encoding
-        self.action_encoding_dim = self.action_space_size
+        self.continuous_action_space = False
+        # The dim of action space. For discrete action space, it is 1.
+        # For continuous action space, it is the dimension of continuous action.
+        self.action_space_dim = action_space_size if self.continuous_action_space else 1
+        assert discrete_action_encoding_type in ['one_hot', 'not_one_hot'], discrete_action_encoding_type
+        self.discrete_action_encoding_type = discrete_action_encoding_type
+        if self.continuous_action_space:
+            self.action_encoding_dim = action_space_size
+        else:
+            if self.discrete_action_encoding_type == 'one_hot':
+                self.action_encoding_dim = action_space_size
+            elif self.discrete_action_encoding_type == 'not_one_hot':
+                self.action_encoding_dim = 1
+
         self.latent_state_dim = latent_state_dim
         self.proj_hid = proj_hid
         self.proj_out = proj_out
@@ -87,7 +101,7 @@ class MuZeroModelMLP(nn.Module):
         )
 
         self.dynamics_network = DynamicsNetwork(
-            action_space_size=action_space_size,
+            action_encoding_dim=self.action_encoding_dim,
             num_channels=self.latent_state_dim + self.action_encoding_dim,
             common_layer_num=2,
             fc_reward_layers=fc_reward_layers,
@@ -198,7 +212,7 @@ class MuZeroModelMLP(nn.Module):
             latent_state = renormalize(latent_state)
         return latent_state
 
-    def _prediction(self, latent_state: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+    def _prediction(self, latent_state: torch.Tensor) ->  Tuple[torch.Tensor, torch.Tensor]:
         """
         Overview:
             Use the representation network to encode the observations into latent state.
@@ -237,21 +251,30 @@ class MuZeroModelMLP(nn.Module):
         # NOTE: the discrete action encoding type is important for some environments
 
         # discrete action space
-        # Stack latent_state with a game specific one hot encoded action
-        if len(action.shape) == 1:
-            # (batch_size, ) -> (batch_size, 1)
-            # e.g.,  torch.Size([8]) ->  torch.Size([8, 1])
-            action = action.unsqueeze(-1)
+        if self.discrete_action_encoding_type == 'one_hot':
+            # Stack latent_state with the one hot encoded action
+            if len(action.shape) == 1:
+                # (batch_size, ) -> (batch_size, 1)
+                # e.g.,  torch.Size([8]) ->  torch.Size([8, 1])
+                action = action.unsqueeze(-1)
 
-        # transform action to one-hot encoding.
-        # action_one_hot shape: (batch_size, action_space_size), e.g., (8, 4)
-        action_one_hot = torch.zeros(action.shape[0], self.action_space_size, device=action.device)
-        # transform action to torch.int64
-        action = action.long()
-        action_one_hot.scatter_(1, action, 1)
-        action_encoding = action_one_hot
+            # transform action to one-hot encoding.
+            # action_one_hot shape: (batch_size, action_space_size), e.g., (8, 4)
+            action_one_hot = torch.zeros(action.shape[0], self.action_space_size, device=action.device)
+            # transform action to torch.int64
+            action = action.long()
+            action_one_hot.scatter_(1, action, 1)
+            action_encoding = action_one_hot
+        elif self.discrete_action_encoding_type == 'not_one_hot':
+            action_encoding = action / self.action_space_size
+            if len(action_encoding.shape) == 1:
+                # (batch_size, ) -> (batch_size, 1)
+                # e.g.,  torch.Size([8]) ->  torch.Size([8, 1])
+                action_encoding = action_encoding.unsqueeze(-1)
 
-        # state_action_encoding shape: (batch_size, latent_state[1] + action_dim])
+        action_encoding = action_encoding.to(latent_state.device).float()
+        # state_action_encoding shape: (batch_size, latent_state[1] + action_dim]) or
+        # (batch_size, latent_state[1] + action_space_size]) depending on the discrete_action_encoding_type.
         state_action_encoding = torch.cat((latent_state, action_encoding), dim=1)
 
         next_latent_state, reward = self.dynamics_network(state_action_encoding)
@@ -296,14 +319,14 @@ class MuZeroModelMLP(nn.Module):
 class DynamicsNetwork(nn.Module):
 
     def __init__(
-            self,
-            action_space_size: int = 2,
-            num_channels: int = 64,
-            common_layer_num: int = 2,
-            fc_reward_layers: SequenceType = [32],
-            output_support_size: int = 601,
-            last_linear_layer_init_zero: bool = True,
-            activation: Optional[nn.Module] = nn.ReLU(inplace=True),
+        self,
+        action_encoding_dim: int = 2,
+        num_channels: int = 64,
+        common_layer_num: int = 2,
+        fc_reward_layers: SequenceType = [32],
+        output_support_size: int = 601,
+        last_linear_layer_init_zero: bool = True,
+        activation: Optional[nn.Module] = nn.ReLU(inplace=True),
     ):
         """
         Overview:
@@ -311,9 +334,7 @@ class DynamicsNetwork(nn.Module):
             reward by the given current latent state and action.
             The networks are mainly build on fully connected layers.
         Arguments:
-            - action_space_size: (:obj:`int`): Action space size, usually an integer number. For discrete action \
-                space, it is the number of discrete actions. For continuous action space, it is the dimension of \
-                continuous action.
+            - action_encoding_dim (:obj:`int`): The dimension of action encoding.
             - num_channels (:obj:`int`): The num of channels in latent states.
             - common_layer_num (:obj:`int`): The number of common layers in dynamics network.
             - fc_reward_layers (:obj:`SequenceType`): The number of hidden layers of the reward head (MLP head).
@@ -324,8 +345,7 @@ class DynamicsNetwork(nn.Module):
         """
         super().__init__()
         self.num_channels = num_channels
-        # for discrete action space, we use one-hot encoding
-        self.action_encoding_dim = action_space_size
+        self.action_encoding_dim = action_encoding_dim
         self.latent_state_dim = self.num_channels - self.action_encoding_dim
 
         self.fc_dynamics = MLP(
@@ -335,7 +355,9 @@ class DynamicsNetwork(nn.Module):
             out_channels=self.latent_state_dim,
             activation=activation,
             norm_type='BN',
-            output_activation=nn.Identity(),
+            output_activation=False,
+            # TODO(pu): check
+            output_norm=True,
             last_linear_layer_init_zero=last_linear_layer_init_zero
         )
 
@@ -346,7 +368,9 @@ class DynamicsNetwork(nn.Module):
             out_channels=output_support_size,
             activation=activation,
             norm_type='BN',
-            output_activation=nn.Identity(),
+            output_activation=False,
+            # TODO(pu): check
+            output_norm=True,
             last_linear_layer_init_zero=last_linear_layer_init_zero
         )
