+++ zoo/box2d/lunarlander/config/lunarlander_disc_muzero_config.py
@@ -34,6 +34,7 @@ lunarlander_muzero_config = dict(
             lstm_hidden_size=256,
             latent_state_dim=256,
             self_supervised_learning_loss=True,  # NOTE: default is False.
+            discrete_action_encoding_type='one_hot',
         ),
         cuda=True,
         env_type='not_board_games',
